StatDeck/Windows/StatDeck.Service/requirements.txt
psutil==5.9.8
pyserial==3.5
pyautogui==0.9.54
# GPU monitoring (NVIDIA)
nvidia-ml-py3==12.535.133
